[stdlib] Source compatibility fix for lazy filter-filter and map-map cases
In order to maintain source compatibility the type-folding behavior
should not be the default in Swift 3 and Swift 4 modes.
